Add a bunch of decorators
Most of these look terrible, so I'll need to put in some effort to
account for the actual heights of characters being decorated. I
was hoping to avoid that, but it now looks necessary. Furthermore,
experiments show that Prima doesn't report the actual height of
rendered glyphs, so I'll have to tabulate it by hand. :-)