Get trickier font faces working, including mathit, mathcal, etc.
It turns out that the trickiest was mathit, because unicode does
not support italic numbers. I had to create a special-case render
subref for italic digits which would change the font face, render
the digit, and change back. I'm not 100% sure if this handles
spacing correctly or not, but at least it renders.